Refactor frame label method
